refactor(website): extract the mobile nav drawer into its own component
Task 7 of the navbar redesign plan: a pure move, no behavior changes.
NavMobile.tsx now owns the hamburger trigger, the focus trap, scroll
lock, inert management, desktop-breakpoint auto-close, focus restore
via requestAnimationFrame, and the Cmd+K search handoff — verbatim
from Nav.tsx, with `open`/`mobileTab` state and refs localized to the
new component. Nav.tsx keeps the <nav> shell, path parsing, navRef,
useNavSurface, and NavDesktop, and passes down isDocsPage,
docsLibrary, activeSection, activeSlug, and navRef.

The one deliberate addition: since NavMobile is invoked once from
inside nav's flex row (where the hamburger button must stay, for
flex layout), the overlay dialog is rendered via createPortal to
document.body so it remains a sibling of <nav> in the DOM rather than
becoming a second direct <div> child of <nav> — preserving both the
original stacking-context fix and immunity from the Task 6
`.nav-bar[data-route='docs'] > div` padding rule. No props beyond the
given NavMobileProps interface were needed for this.

Verified: nx test website -- --run -> 140 files / 1413 tests passed,
identical to baseline. nx build website exits 0. nx lint website ->
0 errors, 65 pre-existing warnings (no new ones). nx e2e website
--testFiles=e2e/nav-height.spec.ts -> 16/16 passed (port 4300 was
free).
